horus: Export `horus_utils:should_process_module/1`
... instead of trying to call it whenever a `should_process_function`
call back is defined.

[Why]
The internal `should_process_module/1` and the callback may take
different decisions. This could lead to unexpected behavior.

[How]
It's simpler to just expose the previously internal function and let the
callback use it if it needs. This way, it can take into account the
default decision or ignore it.

Note that the function was renamed from
`should_collect_code_for_module/1' to `should_process_module/1`.