Add support for cover-compiled modules
[Why]
Extracting a function from a cover-compiled module always worked.
However, we missed the instructions added by `cover` because we took the
module beam from the code server.

This means that executing a standalone function would not emit
coverage information and thus the original source code would appear to
never run.

[How]
In this commit, if the module where the function to extract is defined
is cover-compiled, we take the module beam from the ETS table where
`cover` stores the cover-compiled modules.

This way, we ensure that the generated standalone function will contain
any instructions added by `cover` and coverage information will be
reported as expected.

Fixes #3.