Reconstruct meaningful stacktraces when a standalone fun raises an exception
[Why]
When a standalone function throws an exception, the stacktrace already
shows the correct source file name and line number. However the function
name makes no sense because it is a generated name. Thus it corresponds
to no function in the indicated source file. This makes debugging more
difficult.

[How]
The returned standalone function now contains a mapping from the
generated function name to its original MFA.

In `horus:exec/2`, we then catch the exception and reconstruct the
actual stacktrace using this mapping.

Fixes #2.