Support argument lists in stacktraces
Sometimes, a stacktrace frame will contain the actual list of arguments
to a function instead of just its arity. We need to transfer this list
of arguments to the reconstructed stacktrace.

We take a sublist of the arguments correpsonding to the arity of the
original function. This is required because the extracted standalone
function may take more arguments to take its environment into account.
Arguments from the environment are therefore discarded.