Support cover compiled modules on remote nodes
[Why]
When using `cover` accross several nodes (using `cover:start/1`),
modules are cover-compiled on one node (the "main node") and sent to the
code server instances on other nodes. I.e. only the main node keeps a
copy of the cover-compiled node in an ETS table.

Our initial use of `cover` didn't take this into account. In particular,
`cover:is_compile/1` would return an error we never matched, leading to
a case clause exception.

[How]
We use the undocumented `cover:get_main_node/0` function to learn about
the main node name and RPC call it to get its copy of the cover-compiled
module code.

We can't call the local code server for that because once a module is
loaded, is discards the initial beam binary.