horus: Always allow to override object code
This remains an internal function and is NOT part of the public API.

[Why]
We need this as part of the support of cover-compiled module. In a
future commit, `horus_utils` will override the object code of a module
which won't have a beam file on disk (it will be compiled in memory at
runtime).

Therefore `code:get_object_code()` can't work because it only knows
about beam files on disk. Once loaded, the beam content is discarded and
can't be recovered.

[How]
We simply export the functions which were exported only for tests
before.

At the same time, we add a `forget_overridden_object_code()` function to
clear the overridden object code.