Add special pass1 processing for cover-compiled modules
[Why]
Cover-compiled modules have instructions added by `cover` to update the
executions counters. As of this commit in Erlang 25, a cover-compiled
module does the following two calls for each executed line:
1. it calls `persistent_term:get()` to get the reference to a counter
2. it calls `counters:add()` to bump that counter

The added instructions and these two calls in particular might be denied
by the `ensure_instruction_is_permitted` and `should_process_function`
callbacks provided by the caller. For example, this will be the case of
Khepri which forbids uses of `persistent_term` and `counters`.

Therefore we need a way to skip the instructions added by `cover`. In
other words, we must not use the callbacks for these instructions and
calls, but treat other instructions as usual.

[How]
`horus_cover` uses a very simple `horus_cover_helper` module with a
single function returning `ok` in it. This module is compiled at runtime
with both `compile` and `cover`. It then extracts the two copies of the
function using Horus and compares the assembly. This way, `horus_utils`
isolates the instructions added by `cover`.

The isolated instructions are converted to ETS match spec patterns.

When `horus` processes a cover-compiled module (its object code comes
from `cover`), `horus_cover` compares the instructions to the created
ETS match spec patterns. The matching instructions are not passed to the
callbacks and always added to the result.