erlang_records: Confuse the compiler to avoid move/2 instructions
In `create_record_test/0`, the compiler sees the whole record as a
literal, so it uses a `move/2` instruction with the literal value.
We can confuse it into using a `put_tuple2/2` instruction by masking
the field's value.

In the `update_record_test/0` case, the compiler is smart enough to
notice that update replaces the entire record definition, so it uses
a `move/2` instruction rather than `update_record/5`. We can confuse
the compiler into using an `update_record/5` instruction by using a
record with multiple fields and only updating one field.