test/cached_extractions.erl: Ignore compile:forms/2-related Dialyzer errors
[Why]
What I considered a fix to the `compile:forms/2` function was accepted
in erlang/otp#5293 and available in Erlang/OTP 25.0+. However, this
change was reverted in Erlang/OTP 27.0. I don't know the reason as of
this writing.

Dialyzer in Erlang/OTP 27.0 also improved and now, it is not pleased by
the way we use `compile:forms/2`, directly passing assembler.

[How]
Let's ignore this Dialyzer error, like we ignore it already in `horus`.