Support original assembly as input
[Why]
The default method is to take the beam from the Code server, then to
disassemble it. This has some limitations though: `beam_disasm` does not
return assembly that compiles out-of-the-box. It needs to be patched to
fix errors or unfinished disassembly.

But more importantly in the context of a Khepri cluster, the assembly
and thus the generated standalone function might not load work on
another Erlang node that runs an older version of Erlang. Indeed, that
older version may not understand newer instrucions emitted by a newer
compiler.

[How]
A way to fix this is to continue to compile the modules with the latest
and greated compiler, but to also compile them again to assembly with
the oldest compiler we want to support and make this assembly available
alongside the beam files.

This is what this change allows: Horus will look up a `.S` file in the
code path with the same name as the module it wants to extract code
from. If this file is available, it will use the assembly directly as
its input. Otherwise, it will get the beam and disassemble it as usual.

The instruction processing was extended to handle variants of some
instructions. For instance, call instructions may use an entry label
instead of af MFA: we have to translate that label to an MFA and process
the call. `beam_disasm' should have probably emitted these variants in
the first place.

Note that this commit does not address the changes required to the build
systems of users of Horus (e.g. Khepri and RabbitMQ) to make it able to
use a compiler from a different Erlang install or generate the assembly
file.